same reference file skips database mapping

+++ src/hyb2/mapping/make_hyb_db_2.py
@@ -18,15 +18,17 @@ def make_hyb_db_2(in_fasta):
         raise FileNotFoundError(f"error: {in_fasta} file not found")
 
     prefix = in_fasta.replace(".fasta", "", 1)
-
-    with spinner("making database ") if quiet else contextlib.nullcontext():
-        if not quiet:
-            print("making database:")
-        subprocess.run(["bowtie2-build", in_fasta, prefix], check=True, 
-                       stdout=subprocess.DEVNULL if quiet else None,
-                       stderr=subprocess.DEVNULL if quiet else None
-                       )
-
     tab = in_fasta.replace("fasta", "tab", 1)
-    with open(in_fasta) as fin:
-        Path(tab).write_text(fasta_to_tab(fin.read()))
+
+    if not (Path(f"{prefix}.rev.2.bt2").is_file() or Path(f"{prefix}.rev.2.bt2l").is_file()):
+
+        with spinner("making database ") if quiet else contextlib.nullcontext():
+            if not quiet:
+                print("making database:")
+            subprocess.run(["bowtie2-build", in_fasta, prefix], check=True, 
+                        stdout=subprocess.DEVNULL if quiet else None,
+                        stderr=subprocess.DEVNULL if quiet else None
+                        )
+    if not Path(tab).is_file():
+        with open(in_fasta) as fin:
+            Path(tab).write_text(fasta_to_tab(fin.read()))
